% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try3.R
\name{MyCorr.plot}
\alias{MyCorr.plot}
\title{Visualize correlation between two numerical variables}
\usage{
MyCorr.plot(x, y)
}
\arguments{
\item{x}{numerical vector of the 1st var}

\item{y}{numerical vector of the 2nd var}
}
\description{
Visualize correlation between two numerical variables by scatter plot, with Spearman correlation statistics shown in title
}
