% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try3.R
\name{MyCorr}
\alias{MyCorr}
\title{Compute correlation between two numerical variables}
\usage{
MyCorr(x, y)
}
\arguments{
\item{x}{numerical vector of the 1st var}

\item{y}{numerical vector of the 2nd var}
}
\description{
Compute correlation between two numerical variables and return a "htest" object, the method is Spearman correlation
}
